+++ src/components/pages/User/UserOrder.tsx
@@ -1,0 +1,5 @@
+function UserOrder() {
+  return null;
+}
+
+export default UserOrder;
